Move DeleteContactModal to modal folder. Update getElementByText method.
