+++ e2e/selectors/Modals/DeleteContactModal.selectors.js
@@ -1,7 +1,7 @@
-import messages from '../../../../locales/languages/en.json';
+import messages from '../../../locales/languages/en.json';
 
 // eslint-disable-next-line import/prefer-default-export
-export const DeleteAccountModalSelectorsText = {
+export const DeleteContactModalSelectorsText = {
   MODAL_TITLE: messages.address_book.delete_contact,
   DELETE_BUTTON: messages.address_book.delete,
   CANCEL_BUTTON: messages.address_book.cancel,
